requirements.txt
Flask>=2.0
SQLAlchemy>=1.4
mysql-connector-python>=8.0
redis>=4.0
python-dotenv>=1.0
pytest>=7.0
